svm-conformance: move instr harness
