+++ svm-conformance/Cargo.toml
@@ -35,6 +35,7 @@ cli = ["dep:clap"]
 dev-context-only-utils = []
 dummy-for-ci-check = []
 ffi = [
+    "dep:agave-precompiles",
     "dep:prost",
     "dep:protosol",
     "dep:solana-compute-budget",
@@ -47,6 +48,7 @@ ffi = [
 frozen-abi = []
 
 [dependencies]
+agave-precompiles = { workspace = true, optional = true }
 clap = { version = "4.5.59", features = ["derive"], optional = true }
 prost = { version = "0.11", optional = true }
 protosol = { version = "=10.1.0", optional = true }
@@ -58,8 +60,13 @@ solana-syscalls = { workspace = true, optional = true }
 
 [dev-dependencies]
 bincode = { workspace = true }
+serde = { workspace = true }
+solana-account = { workspace = true }
+solana-clock = { workspace = true }
 solana-rent = { workspace = true }
 solana-sdk-ids = { workspace = true }
+solana-system-interface = { workspace = true }
+solana-system-program = { workspace = true }
 
 [lints]
 workspace = true
